refactor[#155] : marking 관련 api 파일들 이름 수정
