+++ src/features/marking/api/index.ts
@@ -1,6 +1,6 @@
-export * from "./search";
-export * from "./addMarking";
-export * from "./addTempMarking";
+export * from "./getMarkingList";
+export * from "./postMarkingForm";
+export * from "./postTempMarkingForm";
 export * from "./deleteMarking";
 export * from "./likeMarking";
 export * from "./saveMarking";
